backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Werkzeug==2.3.7
PyJWT==2.8.0
ReportLab==4.0.4
PyPDF2==3.0.1
qrcode==7.4.2
Pillow>=10.4.0
python-dateutil==2.8.2
